requirements.txt
requests
gspread
oauth2client
gradio
python-dotenv
google-genai
python-dotenv
